Refactor guides codec files
