+++ apps/nextjs-website/src/lib/cmsApi.ts
@@ -34,9 +34,9 @@ import { makeApiDataProps } from './apiDataPages';
 import { fetchApiData } from './strapi/codecs/ApiDataCodec';
 import { fetchProducts } from './strapi/codecs/ProductCodec';
 import { makeProductsProps } from './products';
-import { fetchGuideList } from './strapi/guideListCodec';
+import { fetchGuideList } from './strapi/codecs/GuideListPagesCodec';
 import { makeGuideListPagesProps } from './guideListPages';
-import { fetchGuides } from './strapi/guidesCodec';
+import { fetchGuides } from './strapi/fetches/fetchGuides';
 import { makeGuidesProps } from './guides';
 import { makeGuide } from '@/_contents/makeDocs';
 import { fetchOverviews } from '@/lib/strapi/overviewsCodec';
